Fix build error when glib isn't new enough to build example browser
I was not generating the makefile for the browser example if the user's glib
wasn't new enough, but I was not removing the directory from SUBDIRS, so it
would go into that directory and try to 'make all' and fail.  This should now be
fixed.

Also fixed another error discovered via 'make dist' in the reference
documentation not being installed correctly